[Decomp][Player] Extract GlyphMgr from Player.cpp
